+++ Source/Frontend/UI/Extensions/ControlExtensions.cs
@@ -1,0 +1,65 @@
+namespace RTCV.UI.Extensions
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Drawing;
+    using System.Runtime.InteropServices;
+    using System.Windows.Forms;
+
+    public static class ControlExtensions
+    {
+        private static NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
+
+        private const int SRCCOPY = 0xCC0020;
+
+        [DllImport("gdi32.dll")]
+        private static extern int BitBlt(IntPtr hdc, int x, int y, int cx, int cy, IntPtr hdcSrc, int x1, int y1, int rop);
+
+        public static Bitmap getFormScreenShot(this Control con)
+        {
+            logger.Trace($"getFormScreenShot ClientRectangle | Width: {con.ClientRectangle.Width} | Height: {con.ClientRectangle.Height} | X: {con.ClientRectangle.X} | Y: {con.ClientRectangle.Y}");
+            try
+            {
+                var bmp = new Bitmap(con.ClientRectangle.Width, con.ClientRectangle.Height);
+                using (var bmpGraphics = Graphics.FromImage(bmp))
+                {
+                    var bmpDC = bmpGraphics.GetHdc();
+                    using (var formGraphics = Graphics.FromHwnd(con.Handle))
+                    {
+                        var formDC = formGraphics.GetHdc();
+                        BitBlt(bmpDC, 0, 0, con.ClientRectangle.Width, con.ClientRectangle.Height, formDC, 0, 0, SRCCOPY);
+                        formGraphics.ReleaseHdc(formDC);
+                    }
+
+                    bmpGraphics.ReleaseHdc(bmpDC);
+                }
+                return bmp;
+            }
+            catch (Exception ex)
+            {
+                logger.Error(ex, $"Failed to get form screenshot.");
+                return new Bitmap(1, 1);
+            }
+        }
+
+        public static List<Control> getControlsWithTag(this Control.ControlCollection controls)
+        {
+            var allControls = new List<Control>();
+
+            foreach (Control c in controls)
+            {
+                if (c.Tag != null)
+                {
+                    allControls.Add(c);
+                }
+
+                if (c.HasChildren)
+                {
+                    allControls.AddRange(c.Controls.getControlsWithTag()); //Recursively check all children controls as well; ie groupboxes or tabpages
+                }
+            }
+
+            return allControls;
+        }
+    }
+}
