+++ Source/Frontend/UI/Modular/UI_ShadowPanel.cs
@@ -3,6 +3,7 @@
     using System;
     using System.Drawing;
     using System.Windows.Forms;
+    using RTCV.UI.Extensions;
 
     #pragma warning disable CA2213 //Component designer classes generate their own Dispose method
     public partial class UI_ShadowPanel : Form
