+++ Source/Frontend/UI/UI_Extensions.cs
@@ -7,18 +7,11 @@
 namespace RTCV.UI
 {
     using System;
-    using System.Collections.Generic;
-    using System.ComponentModel;
-    using System.Diagnostics;
     using System.Drawing;
-    using System.Globalization;
-    using System.Runtime.InteropServices;
     using System.Windows.Forms;
 
     public static class UI_Extensions
     {
-        private static NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
-
         public static DialogResult GetInputBox(string title, string promptText, ref string value)
         {
             var form = new Form();
@@ -80,62 +73,6 @@ namespace RTCV.UI
                 g.FillRectangle(darkBrush, rectSize);
             }
         }
-
-        private const int SRCCOPY = 0xCC0020;
-
-        [DllImport("gdi32.dll")]
-        private static extern int BitBlt(IntPtr hdc, int x, int y, int cx, int cy, IntPtr hdcSrc, int x1, int y1, int rop);
-
-        public static Bitmap getFormScreenShot(this Control con)
-        {
-            logger.Trace($"getFormScreenShot ClientRectangle | Width: {con.ClientRectangle.Width} | Height: {con.ClientRectangle.Height} | X: {con.ClientRectangle.X} | Y: {con.ClientRectangle.Y}");
-            try
-            {
-                var bmp = new Bitmap(con.ClientRectangle.Width, con.ClientRectangle.Height);
-                using (var bmpGraphics = Graphics.FromImage(bmp))
-                {
-                    var bmpDC = bmpGraphics.GetHdc();
-                    using (var formGraphics = Graphics.FromHwnd(con.Handle))
-                    {
-                        var formDC = formGraphics.GetHdc();
-                        BitBlt(bmpDC, 0, 0, con.ClientRectangle.Width, con.ClientRectangle.Height, formDC, 0, 0, SRCCOPY);
-                        formGraphics.ReleaseHdc(formDC);
-                    }
-
-                    bmpGraphics.ReleaseHdc(bmpDC);
-                }
-                return bmp;
-            }
-            catch (Exception ex)
-            {
-                logger.Error(ex, $"Failed to get form screenshot.");
-                return new Bitmap(1, 1);
-            }
-        }
-
-        #region CONTROL EXTENSIONS
-
-        public static List<Control> getControlsWithTag(this Control.ControlCollection controls)
-        {
-            var allControls = new List<Control>();
-
-            foreach (Control c in controls)
-            {
-                if (c.Tag != null)
-                {
-                    allControls.Add(c);
-                }
-
-                if (c.HasChildren)
-                {
-                    allControls.AddRange(c.Controls.getControlsWithTag()); //Recursively check all children controls as well; ie groupboxes or tabpages
-                }
-            }
-
-            return allControls;
-        }
-
-        #endregion CONTROL EXTENSIONS
 
         public class RTC_Standalone_Form : Form { }
     }
